+++ src/Http/Controllers/NotificationController.php
@@ -112,7 +112,7 @@ class NotificationController extends Controller
 
         $notification->markAsRead();
 
-        return new IlluminateJsonResponse(['data' => $this->serialize($notification->fresh())]);
+        return new IlluminateJsonResponse(['data' => $this->serialize($notification)]);
     }
 
     /**